+++ R/epipredict-package.R
@@ -3,6 +3,8 @@
 #' @importFrom rlang := !! %||% as_function global_env set_names !!!
 #' @importFrom rlang is_logical is_true inject enquo enquos expr
 #' @importFrom stats poly predict lm residuals quantile
+#' @importFrom dplyr arrange across all_of any_of bind_rows group_by summarise
+#'             filter mutate select left_join
 #' @importFrom cli cli_abort
 #' @importFrom checkmate assert assert_character assert_int assert_scalar
 #'             assert_logical assert_numeric assert_number assert_integer