+++ R/layer_cdc_flatline_quantiles.R
@@ -170,7 +170,7 @@ slather.layer_cdc_flatline_quantiles <-
       )
     }
     p <- components$predictions
-    ek <- kill_time_value(epi_keys_mold(components$mold))
+    ek <- kill_time_value(key_colnames(components$mold))
     r <- grab_residuals(the_fit, components)
 
     avail_grps <- character(0L)