+++ R/utils-misc.R
@@ -39,9 +39,9 @@ grab_forged_keys <- function(forged, mold, new_data) {
   # 1. these are the keys in the test data after prep/bake
   new_keys <- names(extras)
   # 2. these are the keys in the training data
-  old_keys <- epi_keys_mold(mold)
+  old_keys <- key_colnames(mold)
   # 3. these are the keys in the test data as input
-  new_df_keys <- epi_keys(new_data, extra_keys = setdiff(new_keys, keys[1:2]))
+  new_df_keys <- key_colnames(new_data, extra_keys = setdiff(new_keys, keys[1:2]))
   if (!(setequal(old_keys, new_df_keys) && setequal(new_keys, new_df_keys))) {
     cli::cli_warn(c(
       "Not all epi keys that were present in the training data are available",