+++ tests/testthat/test-epi_recipe.R
@@ -128,7 +128,7 @@ test_that("add/update/adjust/remove epi_recipe works as intended", {
   wf <- epi_workflow() %>%
     add_epi_recipe(r)
 
-  steps <- extract_preprocessor(wf)$steps
+  steps <- workflows::extract_preprocessor(wf)$steps
   expect_equal(length(steps), 3)
   expect_equal(class(steps[[1]]), c("step_epi_lag", "step"))
   expect_equal(steps[[1]]$lag, c(0, 7, 14))
@@ -143,7 +143,7 @@ test_that("add/update/adjust/remove epi_recipe works as intended", {
 
   wf <- update_epi_recipe(wf, r2)
 
-  steps <- extract_preprocessor(wf)$steps
+  steps <- workflows::extract_preprocessor(wf)$steps
   expect_equal(length(steps), 2)
   expect_equal(class(steps[[1]]), c("step_epi_lag", "step"))
   expect_equal(steps[[1]]$lag, c(0, 1))
@@ -152,7 +152,7 @@ test_that("add/update/adjust/remove epi_recipe works as intended", {
 
   # adjust_epi_recipe using step number
   wf <- adjust_epi_recipe(wf, which_step = 2, ahead = 7)
-  steps <- extract_preprocessor(wf)$steps
+  steps <- workflows::extract_preprocessor(wf)$steps
   expect_equal(length(steps), 2)
   expect_equal(class(steps[[1]]), c("step_epi_lag", "step"))
   expect_equal(steps[[1]]$lag, c(0, 1))
@@ -161,7 +161,7 @@ test_that("add/update/adjust/remove epi_recipe works as intended", {
 
   # adjust_epi_recipe using step name
   wf <- adjust_epi_recipe(wf, which_step = "step_epi_ahead", ahead = 8)
-  steps <- extract_preprocessor(wf)$steps
+  steps <- workflows::extract_preprocessor(wf)$steps
   expect_equal(length(steps), 2)
   expect_equal(class(steps[[1]]), c("step_epi_lag", "step"))
   expect_equal(steps[[1]]$lag, c(0, 1))
@@ -170,6 +170,6 @@ test_that("add/update/adjust/remove epi_recipe works as intended", {
 
 
   wf <- remove_epi_recipe(wf)
-  expect_error(extract_preprocessor(wf)$steps)
+  expect_error(workflows::extract_preprocessor(wf)$steps)
   expect_equal(wf$pre$actions$recipe$recipe, NULL)
 })